codebase_genius/backend/requirements.txt
google-generativeai>=0.5.0
gitpython>=3.1.40
tree_sitter>=0.20.2
markdown>=3.5
networkx>=3.2
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
requests>=2.31.0
jaseci>=1.5.0
byllm>=0.2.0